requirements.txt
pandas
matplotlib
streamlit
pyarrow
fastparquet
requests
python-dotenv
google-generativeai